moving tasks into cc_engine
